more monitors into contexts
