+++ council/agents/agent.py
@@ -1,15 +1,13 @@
 import logging
-from typing import Optional, List
+from typing import List, Optional
 
 from council.chains import Chain
-from council.contexts import AgentContext, ChainContext, Budget, InfiniteBudget
-from council.controllers import ControllerBase, BasicController, ExecutionUnit
+from council.contexts import AgentContext, Budget, ChainContext, InfiniteBudget, Monitorable, Monitored
+from council.controllers import BasicController, ControllerBase, ExecutionUnit
 from council.evaluators import BasicEvaluator, EvaluatorBase
-from council.filters import FilterBase, BasicFilter
+from council.filters import BasicFilter, FilterBase
 from council.runners import new_runner_executor
 from council.skills import SkillBase
-from council.monitors import Monitorable, Monitored
-
 from .agent_result import AgentResult
 
 logger = logging.getLogger(__name__)
