+++ council/contexts/__init__.py
@@ -13,5 +13,8 @@ from ._execution_log_entry import ExecutionLogEntry
 from ._llm_context import LLMContext
 from ._message_collection import MessageCollection
 from ._message_list import MessageList
-from ._skill_context import IterationContext, SkillContext
+from ._monitor import Monitor
+from ._monitorable import Monitorable
+from ._monitored import Monitored
 from ._scorer_context import ScorerContext
+from ._skill_context import IterationContext, SkillContext
