+++ council/contexts/_agent_context.py
@@ -1,13 +1,13 @@
 from typing import Iterable, Optional, Sequence
 
-from council.monitors import Monitored
-from ._budget import Budget
 from ._agent_context_store import AgentContextStore
+from ._budget import Budget
 from ._chat_history import ChatHistory
 from ._chat_message import ScoredChatMessage
 from ._context_base import ContextBase
 from ._execution_context import ExecutionContext
 from ._message_collection import MessageCollection
+from ._monitored import Monitored
 
 
 class AgentContext(ContextBase):
