+++ council/contexts/_scorer_context.py
@@ -1,12 +1,11 @@
 from typing import Optional
 
-from council.monitors import Monitored
-from ._budget import Budget, InfiniteBudget
-
 from ._agent_context_store import AgentContextStore
+from ._budget import Budget, InfiniteBudget
 from ._chat_history import ChatHistory
 from ._context_base import ContextBase
 from ._execution_context import ExecutionContext
+from ._monitored import Monitored
 
 
 class ScorerContext(ContextBase):
