+++ council/controllers/controller_base.py
@@ -2,9 +2,8 @@ from abc import ABC, abstractmethod
 from typing import List, Sequence
 
 from council.chains import Chain
-from council.contexts import AgentContext
+from council.contexts import AgentContext, Monitorable
 from .execution_unit import ExecutionUnit
-from council.monitors import Monitorable
 
 
 class ControllerBase(Monitorable, ABC):
