+++ council/controllers/llm_controller.py
@@ -1,14 +1,12 @@
 import logging
-from typing import List, Tuple, Optional
+from typing import List, Optional, Tuple
 
-from council.contexts import AgentContext, LLMContext
 from council.chains import Chain
-from council.llm import LLMMessage, LLMBase
+from council.contexts import AgentContext, LLMContext, Monitored
+from council.llm import LLMBase, LLMMessage
 from council.utils import Option
-
 from .controller_base import ControllerBase
 from .execution_unit import ExecutionUnit
-from ..monitors import Monitored
 
 logger = logging.getLogger(__name__)
