+++ council/llm/llm_base.py
@@ -1,10 +1,9 @@
 import abc
 import logging
+from typing import Any, List, Optional, Sequence
 
-from typing import List, Any, Optional, Sequence
+from council.contexts import Consumption, LLMContext, Monitorable
 from .llm_message import LLMMessage, LLMessageTokenCounterBase
-from council.contexts import LLMContext, Consumption
-from council.monitors import Monitorable
 
 logger = logging.getLogger(__name__)
